Remove obsolete support for nonempty credential without SP
All users with a nonempty lockscreen credential now have a synthetic
password (SP), due to the migration logic that was enabled in
Android O MR1 (commit 7a0cc0a7fbfa / http://ag/2339115).  For this to
not be the case, a device that launched with Android O or earlier would
have to be upgraded all the way to Android U.  That's unsupported; and
even if it was supported, the migration to SP would have happened if the
user was unlocked even just once with *any* version O MR1 through T.
Therefore, there's no longer any need to support the non-SP case except
when credentialType == CREDENTIAL_TYPE_NONE.

Remove all the relevant code that is no longer needed.

Bug: 146040259
Test: atest com.android.server.locksettings
Test: Upgraded a device that has an LSKF set, then unlocked the device
      and changed the LSKF.
Test: m services FrameworksServicesTests RUN_ERROR_PRONE=true
Change-Id: I08ff4d86379e3289ed8aa02b286a92b3ae431367