[Keyguard] Remove behind scrim for alarm
When we show the alarm activity, the behind scrim is showing and
intercepting the touch.

Looking at the scrim state, even though the panel expansion is 0, the
behind scrim is a value greater than 1. This is because we are also
tracking bouncer hidden fraction in the case the scrim state is
UNLOCKED. I believe this case is only relevant to DREAMING. Tracking
bouncer hidden fraction only for DREAMING fixes the issue.

Test: Manual on device plus a unit test
Bug: 232581524
Change-Id: I96542e279c884b313a133ab0a0070b6f0d9e6771