Partial App-ops service refactor.
Added a new Interface for AppOpsService.
Currently this interface has funtions for accessing and modifying
app-ops modes,i.e. package and Uid modes. In future CLs we'll be moving
the callback for watching app-op mode changes to this interface as well
as op restrictions(user and global).

Added a new interface for Scheduling app-ops state persistence, so that
there is only one class for persisting the app-ops state, reducing
complexity.

This CL also includes implementation of the interface, moving the
current implementation of app-ops mode management to the impl.

BUG: 240200048

Test: Manual
Change-Id: I79a48a0afa9489b9281a6b41d92004a3f69e2e95