Deprecate carrier config KEY_SIMPLIFIED_NETWORK_SETTINGS_BOOL
The carrier config key was introduced in Oreo release without
an implementation. It has no behavior impact when carriers
override this carrier config, mismatching what it was delared
in its documentation.

This CC key should be cleared up, not only to keep a healthy
public API surface, but also avoid confusing carriers who
may try to override the CC key to expect a behavior change.

Bug: 242755061
Test: make offline-sdk-docs
Change-Id: Ib30091753eda7dbe30ecc037d4303bf99a4ad3cc