Update FP failure bouncer &  messaging logic
On devices that don't support haptics, when a fingeprint
fails, bring users to the lock screen (instead of bouncer).
Make sure the fingeprint error message will surface once the screen
is on.

In general, show the bouncer on the last attempt that causes a
fingerprint or face lockout event.

Update the UDPFS show-bouncer logic to after 3 consecutive attempts
(instead of 2).

Test: atest BiometricsUnlockControllerTest
Test: atest KeyguardIndicationControllerTest
Fixes: 240381698
Fixes: 242588598
Change-Id: I1ef7be8cdc9abae2cc2f4a9d361fcfb875be1dad