Re-enable secureSettingsBackedUpOrDenied test
This test exists to make sure that every new setting that is introduced
is either backed up or added to the deny list to indicate it shouldn't
be backed up. Over 2 years ago this test was disabled because it appears
to have been failing for some conditions, and unfortunately it was never
re-enabled.

While this test was disabled, many new settings have been introduced
that are not being backed up but are also not present in the deny list.
It will take some time to investigate whether each of these settings
should be backed up or not, but we want to re-enable this test as soon
as possible to avoid this problem with future settings. So I've created
a temporary allowlist of settings that are neither backed up or deny
listed. We should gradually remove the settings from this list as we
figure out whether they should be backed up or not. New settings should
not be added to this list.

Additionally, there was one setting that appeared in both
SETTINGS_TO_BACKUP and BACKUP_DENY_LIST_SECURE_SETTINGS. Since it's
being backed up, I've removed it from the deny list.

Bug: 243630515
Test: atest SettingsBackupTest.java
Change-Id: I399c21ee03041eb0cf7e1c0695bd96edc5e8a08f